# transnexus-vnpt-spoofing/api-docs.json
{"openapi":"3.0.3","info":{"title":"vnptSpoofing API","description":"vnptSpoofing API documentation","contact":{},"license":{"name":"unlicensed"},"version":"0.0.1"},"servers":[{"url":"http://localhost:8087","description":"Inferred Url"}],"tags":[{"name":"blacklist-resource","description":"Blacklist Resource"},{"name":"call-log-resource","description":"Call Log Resource"},{"name":"call-stat-resource","description":"Call Stat Resource"},{"name":"routing-rule-resource","description":"Routing Rule Resource"},{"name":"simulator-resource","description":"Simulator Resource"},{"name":"whitelist-resource","description":"Whitelist Resource"}],"paths":{"/api/blacklists":{"get":{"tags":["blacklist-resource"],"summary":"getAllBlacklists","operationId":"getAllBlacklistsUsingGET","parameters":[{"name":"page","in":"query","description":"Page number of the requested page","required":false,"style":"form","allowReserved":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","description":"Size of a page","required":false,"style":"form","allowReserved":true,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"style":"form","allowReserved":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Blacklist"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["blacklist-resource"],"summary":"updateBlacklist","operationId":"updateBlacklistUsingPUT","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Blacklist"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Blacklist"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["blacklist-resource"],"summary":"createBlacklist","operationId":"createBlacklistUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Blacklist"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Blacklist"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/blacklists/{id}":{"get":{"tags":["blacklist-resource"],"summary":"getBlacklist","operationId":"getBlacklistUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Blacklist"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["blacklist-resource"],"summary":"deleteBlacklist","operationId":"deleteBlacklistUsingDELETE","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/call-logs":{"get":{"tags":["call-log-resource"],"summary":"getAllCallLogs","operationId":"getAllCallLogsUsingGET","parameters":[{"name":"page","in":"query","description":"Page number of the requested page","required":false,"style":"form","allowReserved":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","description":"Size of a page","required":false,"style":"form","allowReserved":true,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"style":"form","allowReserved":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CallLog"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["call-log-resource"],"summary":"updateCallLog","operationId":"updateCallLogUsingPUT","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CallLog"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallLog"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["call-log-resource"],"summary":"createCallLog","operationId":"createCallLogUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CallLog"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallLog"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/call-logs/{id}":{"get":{"tags":["call-log-resource"],"summary":"getCallLog","operationId":"getCallLogUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallLog"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["call-log-resource"],"summary":"deleteCallLog","operationId":"deleteCallLogUsingDELETE","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/daily-report":{"get":{"tags":["call-stat-resource"],"summary":"dailyReport","operationId":"dailyReportUsingGET","parameters":[{"name":"from","in":"query","description":"from","required":true,"style":"form","schema":{"type":"string","format":"date-time"}},{"name":"to","in":"query","description":"to","required":true,"style":"form","schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/daily-stats":{"get":{"tags":["call-stat-resource"],"summary":"dailyStatsByStateIn","operationId":"dailyStatsByStateInUsingGET","parameters":[{"name":"from","in":"query","description":"from","required":true,"style":"form","schema":{"type":"string","format":"date"}},{"name":"to","in":"query","description":"to","required":false,"style":"form","schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/statistics/error-codes":{"get":{"tags":["call-stat-resource"],"summary":"getDistinctErrorCodes","operationId":"getDistinctErrorCodesUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"integer","format":"int32"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/statistics/states":{"get":{"tags":["call-stat-resource"],"summary":"getDistinctState","operationId":"getDistinctStateUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"integer","format":"int32"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/routing-rules":{"get":{"tags":["routing-rule-resource"],"summary":"getAllRoutingRules","operationId":"getAllRoutingRulesUsingGET","parameters":[{"name":"page","in":"query","description":"Page number of the requested page","required":false,"style":"form","allowReserved":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","description":"Size of a page","required":false,"style":"form","allowReserved":true,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"style":"form","allowReserved":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoutingRule"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["routing-rule-resource"],"summary":"updateRoutingRule","operationId":"updateRoutingRuleUsingPUT","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutingRule"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoutingRule"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["routing-rule-resource"],"summary":"createRoutingRule","operationId":"createRoutingRuleUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoutingRule"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoutingRule"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/routing-rules/{id}":{"get":{"tags":["routing-rule-resource"],"summary":"getRoutingRule","operationId":"getRoutingRuleUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoutingRule"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["routing-rule-resource"],"summary":"deleteRoutingRule","operationId":"deleteRoutingRuleUsingDELETE","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/lookup-result":{"post":{"tags":["simulator-resource"],"summary":"fullSimulator","operationId":"fullSimulatorUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CallLog"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallLog"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/lookup-route":{"get":{"tags":["simulator-resource"],"summary":"getBestRoute","operationId":"getBestRouteUsingGET","parameters":[{"name":"msisdn","in":"query","description":"msisdn","required":true,"style":"form","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoutingRule"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/public/perform-spoofing":{"get":{"tags":["simulator-resource"],"summary":"performCallSpoofing","operationId":"performCallSpoofingUsingGET","parameters":[{"name":"calling","in":"query","description":"calling","required":true,"style":"form","schema":{"type":"string"}},{"name":"called","in":"query","description":"called","required":true,"style":"form","schema":{"type":"string"}},{"name":"callref","in":"query","description":"callref","required":false,"style":"form","schema":{"type":"string"}},{"name":"channel","in":"query","description":"channel","required":false,"style":"form","schema":{"type":"string"}},{"name":"peer","in":"query","description":"peer","required":false,"style":"form","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallLog"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/whitelists":{"get":{"tags":["whitelist-resource"],"summary":"getAllWhitelists","operationId":"getAllWhitelistsUsingGET","parameters":[{"name":"page","in":"query","description":"Page number of the requested page","required":false,"style":"form","allowReserved":true,"schema":{"type":"integer","format":"int32"}},{"name":"size","in":"query","description":"Size of a page","required":false,"style":"form","allowReserved":true,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"style":"form","allowReserved":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Whitelist"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["whitelist-resource"],"summary":"updateWhitelist","operationId":"updateWhitelistUsingPUT","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Whitelist"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Whitelist"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["whitelist-resource"],"summary":"createWhitelist","operationId":"createWhitelistUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Whitelist"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Whitelist"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/whitelists/{id}":{"get":{"tags":["whitelist-resource"],"summary":"getWhitelist","operationId":"getWhitelistUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Whitelist"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["whitelist-resource"],"summary":"deleteWhitelist","operationId":"deleteWhitelistUsingDELETE","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"components":{"schemas":{"Blacklist":{"title":"Blacklist","required":["prefix","state"],"type":"object","properties":{"description":{"type":"string"},"id":{"type":"integer","format":"int64"},"length":{"type":"integer","format":"int32"},"prefix":{"type":"integer","format":"int64"},"state":{"type":"integer","format":"int32"}}},"CallLog":{"title":"CallLog","required":["called","calling","createdAt","requestAt","state"],"type":"object","properties":{"blacklisted":{"type":"boolean"},"called":{"type":"string"},"calling":{"type":"string"},"callref":{"type":"string"},"channel":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"dryRun":{"type":"boolean"},"errorCode":{"type":"integer","format":"int32"},"errorDesc":{"type":"string"},"id":{"type":"integer","format":"int64"},"peerIp":{"type":"string"},"requestAt":{"type":"string","format":"date-time"},"responseAt":{"type":"string","format":"date-time"},"route":{"type":"string"},"routeLength":{"type":"integer","format":"int32"},"routePrefix":{"type":"integer","format":"int64"},"sriAt":{"type":"string","format":"date-time"},"sriRespAt":{"type":"string","format":"date-time"},"state":{"type":"integer","format":"int32"},"vlr":{"type":"string"},"whitelisted":{"type":"boolean"}}},"RoutingRule":{"title":"RoutingRule","required":["aprefix","name","state"],"type":"object","properties":{"alength":{"type":"integer","format":"int32"},"aprefix":{"type":"integer","format":"int64"},"description":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"state":{"type":"integer","format":"int32"}}},"Whitelist":{"title":"Whitelist","required":["prefix","state"],"type":"object","properties":{"description":{"type":"string"},"id":{"type":"integer","format":"int64"},"length":{"type":"integer","format":"int32"},"prefix":{"type":"integer","format":"int64"},"state":{"type":"integer","format":"int32"}}}}}}